pyproject.toml
[tool.poetry]
name = "dbgen"
version = "0.1.0"
description = ""
authors = ["syfun <sunyu418@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
sqlalchemy = "^1.3.15"
mako = "^1.1.2"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
